feat(connectors): consolidate Owletto-for-Chrome primitives into one `chrome` connector
Replace four standalone ConnectorDefinitions (`chrome.tabs`,
`browser.evaluate`, `browser.page_text`, `browser.fill_form`) with one
`chrome` connector whose feeds (open_tabs / evaluate / page_text /
fill_form) cover the same capabilities. Each paired Chrome profile is
conceptually one integration; pre-consolidation the admin UI listed
four peer entries under one chrome-extension device, which
misrepresented the shape.

The SDK already had `feeds` and `actions` as first-class fields on
ConnectorDefinition; the four old defs just weren't using that shape.

- packages/connectors/src/chrome.ts: new unified connector. Required
  capability `browser.debugger` (strict union of the underlying needs).
- packages/connectors/src/{chrome_tabs.ts, browser/}: deleted.
- packages/connectors/src/index.ts: export the new connector only.
- db/migrations/20260518030000_consolidate_chrome_connector.sql:
  collapses connections per (org, device_worker_id) tuple (including
  unpinned rows), repoints feeds + pending runs to the canonical row,
  renames feed_keys 'page' → 'page_text' and 'fill' → 'fill_form',
  archives the old connector_definitions, wraps the whole migration in
  a transaction.

Worker dispatch lands in lobu-ai/owletto in a paired PR (Owletto for
Chrome 0.4.0) and keeps a legacy fallback for chrome.tabs / browser.*
to cover the race window between migration and gateway code-cut.

Follow-ups (not in this PR):
- device-reconcile.ts currently reuses connections by (org, connector_key,
  auth_profile_id=NULL) without forcing a per-device split. With one
  unified `chrome` connector this becomes more visible if a user has
  two Chrome profiles in one org. Pre-existing; not regressed by this
  PR.
